finite difference stencil method
