+++ server/logic/derivatives.py
@@ -1,9 +1,10 @@
 import numpy as np
 from scipy.special import legendre
 from scipy.integrate import quad
+import scipy.linalg as sla
 
 
-float_type = np.float32
+float_type = np.float64
 
 
 def choose(n, k):
@@ -26,6 +27,39 @@ def integral(u, f, x, h, n):
     l = leg(u)
     return f(x+h*u)*l
 
+def sampling(n):
+    #2n+1 samples
+    return np.array([i for i in range(-n, n+1)])
+
+def factorial(n):
+    facts = [1]
+    for i in range(1, n + 1):
+        facts.append(facts[-1]*i)
+    return facts[-1]
+
+def find_middle(xs):
+    if np.size(xs) % 2 == 0:
+        return xs[int(np.size(xs)/2)]
+    else:
+        return xs[int((np.size(xs) - 1)/2)]
+
+def finite_difference(f, x, n):
+    samples = sampling(n)
+    mat = np.array([
+        [np.power(samples[i], j) for i in range(len(samples))] for j in range(len(samples))
+    ])
+    
+    b = np.zeros(len(samples)).transpose()
+    b[n] += factorial(n)
+
+    L, U = sla.lu(mat, permute_l=True)
+    y = sla.solve(L, b)
+    v = sla.solve(U, y)
+    if isinstance(x, int):
+        x = np.array([x])
+    mid = find_middle(x)
+ 
+
 def lanczo(f, x, n):
     leg = legendre(n)
     h = np.float_power(10,-3)
@@ -47,3 +81,5 @@ if __name__ == "__main__":
     # debug code
     print(newton(lambda x: (x**4), 2, 3))
     print(lanczo(lambda x: (x**4), 2, 3))
+    print(finite_difference(lambda x: (x**4), 2, 3))
+
